split off trait CommandArguments from ExecOneCommand to be able to use it stand-alone
fixes #105
